requirements.txt
django
djangorestframework
python-dotenv
psycopg[binary]